Extract misaligned load/store emulation
This commit extracts the load/store emulation logic from the protect
payload policy to anticipate the addition of a new offload policy which
will also emulate misaligned load and stores. With the change both
modules will be able to use the same implementation.
